Fix click event of Cards.js
Modified handleClick function to reference 'a.card' node in the
'target' variable. This change is necessary to correctly update the
pokemon state variable since the markup of 'a.card' node changed, but
handleClick's code didn't adapt to those changes.

Additionally, a line has been added to blur the selected anchor to
improve user experience.